requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.2.0
django-notifications-hq==1.7.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonfield==3.1.0
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pymongo==3.12.3
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.2.4
swapper==1.3.0
uritemplate==4.1.1
urllib3==1.26.9
